Remove unnecessary things
